fix all bugs

+++ backend/products/views.py
@@ -104,16 +104,21 @@ class OrderViewSet(viewsets.ModelViewSet):
     @action(detail=False, methods=['get'])
     def claim_order(self, request):
         order_id = request.query_params.get('order_id', None)
-        order = get_object_or_404(Order, order_id=order_id)
-        order.is_delivered = True
-        item = order.item
-        order.to_address = request.query_params.get('to_address')
-        warranty_period = item.product.warranty_period
-        delta = datetime.timedelta(days=warranty_period)
-        item.warranty_end_date = datetime.date.today() + delta
-        if request.query_params.get('nft_id'):
-            item.nft_id = request.query_params.get('nft_id')
-        item.save()
-        order.save()
-        order.delete()
-        return Response(OrderSerializer(order).data)
+        to_address = request.query_params.get('to_address', None)
+        nft_id = request.query_params.get('nft_id', None)
+        print(order_id, to_address, nft_id)
+        if order_id:
+            order = get_object_or_404(Order, order_id=order_id)
+            item = order.item
+            owner, created = Owner.objects.get_or_create(wallet_address=to_address)
+            item.owner = owner
+            if nft_id:
+                item.nft_id = nft_id
+            warranty_period = item.product.warranty_period
+            delta = datetime.timedelta(days=warranty_period)
+            item.warranty_end_date = datetime.date.today() + delta*30
+            item.save()
+            order.save()
+            order.delete()
+            return Response(OrderSerializer(order).data)
+        return Response(status=400)
